Refactor remote/local logic into UserMediaRepository
